backend/requirements.txt
flask
flask-cors
python-dotenv
torch
transformers
requests